Make query count assertion code reusable by merging all test classes into one
