+++ agents/test/BUILD
@@ -1,96 +1,10 @@
-
 java_test(
-    name = "MarriageAgentE2E",
+    name = "QueryCountE2E",
     srcs = [
-        "MarriageAgentE2E.java",
+        "QueryCountE2E.java",
         "common/TestArgsInterpreter.java",
     ],
-    test_class = "grakn.simulation.agents.test.MarriageAgentE2E",
-    deps = [
-        # External dependencies from @graknlabs
-        "@graknlabs_client_java//:client-java",
-        "@graknlabs_graql//java:graql",
-
-        # External dependencies from Maven
-        "//dependencies/maven/artifacts/org/hamcrest:hamcrest",
-    ],
-)
-
-java_test(
-    name = "ParentshipAgentE2E",
-    srcs = [
-        "ParentshipAgentE2E.java",
-        "common/TestArgsInterpreter.java",
-    ],
-    test_class = "grakn.simulation.agents.test.ParentshipAgentE2E",
-    deps = [
-        # External dependencies from @graknlabs
-        "@graknlabs_client_java//:client-java",
-        "@graknlabs_graql//java:graql",
-
-        # External dependencies from Maven
-        "//dependencies/maven/artifacts/org/hamcrest:hamcrest",
-    ],
-)
-
-java_test(
-    name = "RelocationAgentE2E",
-    srcs = [
-        "RelocationAgentE2E.java",
-        "common/TestArgsInterpreter.java",
-    ],
-    test_class = "grakn.simulation.agents.test.RelocationAgentE2E",
-    deps = [
-        # External dependencies from @graknlabs
-        "@graknlabs_client_java//:client-java",
-        "@graknlabs_graql//java:graql",
-
-        # External dependencies from Maven
-        "//dependencies/maven/artifacts/org/hamcrest:hamcrest",
-    ],
-)
-
-java_test(
-    name = "CompanyAgentE2E",
-    srcs = [
-        "CompanyAgentE2E.java",
-        "common/TestArgsInterpreter.java",
-    ],
-    test_class = "grakn.simulation.agents.test.CompanyAgentE2E",
-    deps = [
-        # External dependencies from @graknlabs
-        "@graknlabs_client_java//:client-java",
-        "@graknlabs_graql//java:graql",
-
-        # External dependencies from Maven
-        "//dependencies/maven/artifacts/org/hamcrest:hamcrest",
-    ],
-)
-
-java_test(
-    name = "EmploymentAgentE2E",
-    srcs = [
-        "EmploymentAgentE2E.java",
-        "common/TestArgsInterpreter.java",
-    ],
-    test_class = "grakn.simulation.agents.test.EmploymentAgentE2E",
-    deps = [
-        # External dependencies from @graknlabs
-        "@graknlabs_client_java//:client-java",
-        "@graknlabs_graql//java:graql",
-
-        # External dependencies from Maven
-        "//dependencies/maven/artifacts/org/hamcrest:hamcrest",
-    ],
-)
-
-java_test(
-    name = "ProductAgentE2E",
-    srcs = [
-        "ProductAgentE2E.java",
-        "common/TestArgsInterpreter.java",
-    ],
-    test_class = "grakn.simulation.agents.test.ProductAgentE2E",
+    test_class = "grakn.simulation.agents.test.QueryCountE2E",
     deps = [
         # External dependencies from @graknlabs
         "@graknlabs_client_java//:client-java",
